Job_Portal/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
future==0.18.2
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
ipython

multidict==4.7.6
numpy
oauth2client==4.1.3
oauthlib==3.1.0
pandas
python-dotenv==0.15.0

pytz==2020.1
